GLOB-529: new requirement [1. MARK ZERO TIME has to be default observation if the user is annotating for the first time. 2. Once MARK ZERO TIME is created :  a. A user can't edit MARK_ZERO_TIME observation  b. A User can't delete MARK_ZERO_TIME observation 3.While changing MARK_ZERO_TIME observation  a. User will see a pop if he is trying to create some other observation first time.  b. While delting also user will see a pop up saying MARK ZERO TIME can't be deleted. 4. Rest functionality are same as per requirement]

+++ annotation_view/observation_table.py
@@ -5,7 +5,7 @@ from PyQt4.QtGui import *
 from video_player import DialogActions
 from annotation_view.util import ObservationColumn,ColumnsEnum
 
-
+MARK_ZERO_TIME_ID = 16
 class ObservationTableModel(QAbstractTableModel):
     observationUpdated = pyqtSignal(Observation)
     eventUpdated = pyqtSignal(Event)
@@ -270,7 +270,13 @@ class ObservationTable(QTableView):
     def remove_observation(self, obs):
         self.clearSelection()
         # TODO busy cursor
-        self.current_set.delete_observation(obs)
+        if obs.events[0].attribute[0]['id'] == MARK_ZERO_TIME_ID :
+            msg = 'Mark Zero Observation delete not allowed!!'
+            QMessageBox.question(self, 'Delete confirmation', msg, QMessageBox.Close)
+            self.dialog_values['attribute'] = [MARK_ZERO_TIME_ID]
+        else :
+           self.current_set.delete_observation(obs)
+
         self.refresh_model()
 
     def empty(self):
